+++ rcopula/fit/__init__.py
@@ -3,13 +3,16 @@
 from __future__ import annotations
 
 from rcopula.fit.api import METHODS, fit, loglik_copula, nearest_correlation
+from rcopula.fit.mvdc import JointFitResult, fit_joint
 from rcopula.fit.results import CopulaFitResult
 from rcopula.fit.variance import var_itau, var_ml, var_mpl
 
 __all__ = [
     "METHODS",
     "CopulaFitResult",
+    "JointFitResult",
     "fit",
+    "fit_joint",
     "loglik_copula",
     "nearest_correlation",
     "var_itau",